+++ src/open_inwoner/haalcentraal/api.py
@@ -135,13 +135,13 @@ class BRP_2_1(BRPAPI):
 
         # Centric headers
         if self.config.x_request_organization:
-            headers["x-requets-organization"] = self.config.x_request_organization
+            headers["x-requests-organization"] = self.config.x_request_organization
         if self.config.x_request_application:
-            headers["x-requets-application"] = self.config.x_request_application
+            headers["x-requests-application"] = self.config.x_request_application
         if self.config.x_request_afnemerscode:
-            headers["x-requets-afnemerscode"] = self.config.x_request_afnemerscode
+            headers["x-requests-afnemerscode"] = self.config.x_request_afnemerscode
         if self.config.x_request_user:
-            headers["x-requets-user"] = self.config.x_request_user
+            headers["x-requests-user"] = self.config.x_request_user
 
         response = self.client.post(
             url=url,